move csv to utils
