+++ register/utils.py
@@ -1,3 +1,6 @@
+import csv
+
+from django.http import HttpResponse
 from django.urls import reverse as django_reverse
 
 
@@ -13,3 +16,37 @@ def reverse(viewname, args=None, kwargs=None, request=None, format=None, **extra
     if request:
         return request.build_absolute_uri(url)
     return url
+
+
+# Code inspired by this snippet: https://gist.github.com/mgerring/3645889
+
+def export_as_csv_action(description="Export selected objects as CSV file",
+                         fields=None, exclude=None, header=True):
+    """
+    This function returns an export csv action
+    'fields' and 'exclude' work like in django ModelForm
+    'header' is whether or not to output the column names as the first row
+    """
+
+    def export_as_csv(modeladmin, request, queryset):
+        opts = modeladmin.model._meta
+
+        if not fields:
+            field_names = [field.name for field in opts.fields]
+        else:
+            field_names = fields
+
+        response = HttpResponse()
+        response['Content-Disposition'] = 'attachment; filename=%s.csv' % str(opts).replace('.', '_')
+
+        writer = csv.writer(response)
+        if header:
+            writer.writerow(field_names)
+        for obj in queryset:
+            row = [getattr(obj, field)() if callable(getattr(obj, field)) else getattr(obj, field) for field in
+                   field_names]
+            writer.writerow(row)
+        return response
+
+    export_as_csv.short_description = description
+    return export_as_csv
